update i18 texts & reorganize files
