+++ src/pages/landing/Landing.jsx
@@ -2,8 +2,10 @@ import Carrousel from "../../components/Carrousel";
 import InfoCard from "../../components/InfoCard";
 import testFoto from "../../assets/goldCircuit1.png";
 import ProductCard from "../../components/ProductCard";
+import { useTranslation } from "react-i18next";
 
 export default function Landing() {
+  const { t } = useTranslation();
   return (
     <div className="content h-full ">
       <section className=" flex flex-col gap-[5rem]">
@@ -12,7 +14,7 @@ export default function Landing() {
           <img src={testFoto} className="place-self-end  md:w-[70%] lg:w-[45%] lg:place-self-center" />
         </section>
         <div className="">
-          <h1 className="mb-10 w-fit mx-auto xl:text-3xl">los más vendidos</h1>
+          <h1 className="mx-auto mb-10 w-fit xl:text-3xl">{t('common.top-sellers')}</h1>
           <Carrousel
             content={[
               { component: <ProductCard imgUrl={testFoto} name={"CIRCUIT"} price={"$5.000"} />, qty: 3 },
