+++ src/utils/index.js
@@ -1,11 +1,13 @@
-// En este archivo van funciones utiles que usaremos a lo largo del proyecto
+export const links = [
+  { name: "PRODUCTOS", path: "/products" },
+  { name: "CATEGORIAS", path: "/products/category" },
+  { name: "BLOG", path: "/blog" },
+  { name: "SOBRE NOSOTROS", path: "/aboutUs" },
+];
 
-// export function arrayPaginator(arr, cardsPerPage, page) {
-//   const start = page * cardsPerPage
-//   const end = start + cardsPerPage
-
-//   return {
-//     newArr,
-//     nButtons: arr?.length / cardsPerPage,
-//   }
-// }
+export const categories = [
+  { name: "SISTEMAS", value: 1 },
+  { name: "PLUGS", value: 2 },
+  { name: "SAFE ROOTS", value: 3 },
+  { name: "CANASTAS", value: 4 },
+];
